Run with logback 1.0.13
* Remove the UDP appender as the API changes aren't trivial
* Report event timestamp instead of elapsed time
* Use ints for timeouts instead of longs

+++ src/main/java/net/logstash/logback/appender/LogstashTcpSocketAppender.java
@@ -50,26 +50,26 @@ import ch.qos.logback.core.util.Duration;
  * <br/>
  * For example:<br/>
  * <code>logger.info(ClassicConstants.FINALIZE_SESSION_MARKER, "About to end the job");</code>
- * 
- * 
+ *
+ *
  * @author <a href="mailto:mirko.bernardoni@gmail.com">Mirko Bernardoni</a>
  * @since 11 Jun 2014 (creation date)
  */
 public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
         implements Runnable, SocketConnector.ExceptionHandler {
-    
+
     private static final PreSerializationTransformer<ILoggingEvent> PST = new LoggingEventPreSerializationTransformer();
-    
+
     /**
      * The default port number of remote logging server (4560).
      */
     public static final int DEFAULT_PORT = 4560;
-    
+
     /**
      * The default reconnection delay (30000 milliseconds or 30 seconds).
      */
     public static final int DEFAULT_RECONNECTION_DELAY = 30000;
-    
+
     /**
      * Default size of the queue used to hold logging events that are destined
      * for the remote peer.
@@ -77,50 +77,50 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
      * using about 10MB additional memory if the queue is full
      */
     public static final int DEFAULT_QUEUE_SIZE = 10000;
-    
+
     /**
      * Default timeout when waiting for the remote server to accept our
      * connection.
      */
     private static final int DEFAULT_ACCEPT_CONNECTION_DELAY = 5000;
-    
+
     /**
      * Default timeout for how long to wait when inserting an event into the
      * BlockingQueue.
      */
     private static final int DEFAULT_EVENT_DELAY_TIMEOUT = 100;
-    
+
     private String remoteHost;
-    
+
     private int port = DEFAULT_PORT;
-    
+
     private InetAddress address;
-    
+
     private Duration reconnectionDelay = new Duration(
             DEFAULT_RECONNECTION_DELAY);
-    
+
     private int acceptConnectionTimeout = DEFAULT_ACCEPT_CONNECTION_DELAY;
-    
+
     private Duration eventDelayLimit = new Duration(DEFAULT_EVENT_DELAY_TIMEOUT);
 
     private int queueSize = DEFAULT_QUEUE_SIZE;
-    
+
     private BlockingQueue<ILoggingEvent> queue;
-    
+
     private String peerId;
-    
+
     private Future<?> task;
-    
+
     private Future<Socket> connectorTask;
-    
+
     private volatile Socket socket;
-    
+
     /**
      * It is the encoder which is ultimately responsible for writing the event
      * to an {@link OutputStream}.
      */
     protected Encoder<ILoggingEvent> encoder;
-    
+
     private boolean includeCallerData;
 
     /**
@@ -129,7 +129,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
     public Encoder<ILoggingEvent> getEncoder() {
         return encoder;
     }
-    
+
     /**
      * @param encoder
      *            the encoder to set
@@ -137,17 +137,17 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
     public void setEncoder(Encoder<ILoggingEvent> encoder) {
         this.encoder = encoder;
     }
-    
+
     /**
      * Get the pre-serialization transformer that will be used to transform each
      * event into a Serializable object before delivery to the remote receiver.
-     * 
+     *
      * @return transformer object
      */
     public PreSerializationTransformer<ILoggingEvent> getPST() {
         return PST;
     }
-    
+
     protected void encoderInit(OutputStream outputStream) {
         if (encoder != null && outputStream != null) {
             try {
@@ -160,7 +160,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             }
         }
     }
-    
+
     protected void encoderClose(OutputStream outputStream) {
         if (encoder != null && outputStream != null) {
             try {
@@ -173,7 +173,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             }
         }
     }
-    
+
     /**
      * {@inheritDoc}
      */
@@ -191,14 +191,14 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
                     + name
                     + ". For more information, please visit http://logback.qos.ch/codes.html#socket_no_port");
         }
-        
+
         if (remoteHost == null) {
             errorCount++;
             addError("No remote host was configured for appender "
                     + name
                     + ". For more information, please visit http://logback.qos.ch/codes.html#socket_no_host");
         }
-        
+
         if (errorCount == 0) {
             try {
                 address = InetAddress.getByName(remoteHost);
@@ -207,7 +207,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
                 errorCount++;
             }
         }
-        
+
         if (errorCount == 0) {
             encoder.start();
             queue = new LinkedBlockingQueue<ILoggingEvent>(queueSize);
@@ -216,7 +216,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             super.start();
         }
     }
-    
+
     /**
      * {@inheritDoc}
      */
@@ -230,7 +230,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             connectorTask.cancel(true);
         super.stop();
     }
-    
+
     /**
      * {@inheritDoc}
      */
@@ -238,7 +238,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
     protected void append(ILoggingEvent event) {
         if (event == null || !isStarted())
             return;
-        
+
         try {
             event.prepareForDeferredProcessing();
             if (includeCallerData) {
@@ -255,7 +255,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             addError("Interrupted while appending event to SocketAppender", e);
         }
     }
-    
+
     /**
      * {@inheritDoc}
      */
@@ -264,11 +264,11 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             while (!Thread.currentThread().isInterrupted()) {
                 SocketConnector connector = createConnector(address, port, 0,
                         reconnectionDelay.getMilliseconds());
-                
+
                 connectorTask = activateConnector(connector);
                 if (connectorTask == null)
                     break;
-                
+
                 socket = waitForConnectorToReturnASocket();
                 if (socket == null)
                     break;
@@ -279,7 +279,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
         }
         addInfo("shutting down");
     }
-    
+
     private SocketConnector createConnector(InetAddress address, int port,
             int initialDelay, long retryDelay) {
         SocketConnector connector = newConnector(address, port, initialDelay,
@@ -288,7 +288,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
         connector.setSocketFactory(getSocketFactory());
         return connector;
     }
-    
+
     private Future<Socket> activateConnector(SocketConnector connector) {
         try {
             return getContext().getExecutorService().submit(connector);
@@ -296,7 +296,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             return null;
         }
     }
-    
+
     private Socket waitForConnectorToReturnASocket()
             throws InterruptedException {
         try {
@@ -307,10 +307,10 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             return null;
         }
     }
-    
+
     /**
      * Inifinte loop that send the messages from the queue to the remote host
-     * 
+     *
      * @throws InterruptedException
      */
     private void dispatchEvents() throws InterruptedException {
@@ -344,7 +344,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             addInfo(peerId + "connection closed");
         }
     }
-    
+
     /**
      * {@inheritDoc}
      */
@@ -357,12 +357,12 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
             addInfo(peerId + ex);
         }
     }
-    
+
     /**
      * Creates a new {@link SocketConnector}.
      * <p>
      * The default implementation creates an instance of {@link DefaultSocketConnector}. A subclass may override to provide a different {@link SocketConnector} implementation.
-     * 
+     *
      * @param address
      *            target remote address
      * @param port
@@ -375,10 +375,10 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
      */
     protected SocketConnector newConnector(InetAddress address, int port,
             long initialDelay, long retryDelay) {
-        return new DefaultSocketConnector(address, port, initialDelay,
-                retryDelay);
+        return new DefaultSocketConnector(address, port, (int) initialDelay,
+                (int) retryDelay);
     }
-    
+
     /**
      * Gets the default {@link SocketFactory} for the platform.
      * <p>
@@ -387,7 +387,7 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
     protected SocketFactory getSocketFactory() {
         return SocketFactory.getDefault();
     }
-    
+
     /**
      * The <b>RemoteHost</b> property takes the name of of the host where a
      * corresponding server is running.
@@ -395,14 +395,14 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
     public void setRemoteHost(String host) {
         remoteHost = host;
     }
-    
+
     /**
      * Returns value of the <b>RemoteHost</b> property.
      */
     public String getRemoteHost() {
         return remoteHost;
     }
-    
+
     /**
      * The <b>Port</b> property takes a positive integer representing the port
      * where the server is waiting for connections.
@@ -410,57 +410,57 @@ public class LogstashTcpSocketAppender extends AppenderBase<ILoggingEvent>
     public void setPort(int port) {
         this.port = port;
     }
-    
+
     /**
      * Returns value of the <b>Port</b> property.
      */
     public int getPort() {
         return port;
     }
-    
+
     /**
      * The <b>reconnectionDelay</b> property takes a positive {@link Duration} value representing the time to wait between each failed connection
      * attempt to the server. The default value of this option is to 30 seconds.
-     * 
+     *
      * <p>
      * Setting this option to zero turns off reconnection capability.
      */
     public void setReconnectionDelay(Duration delay) {
         this.reconnectionDelay = delay;
     }
-    
+
     /**
      * Returns value of the <b>reconnectionDelay</b> property.
      */
     public Duration getReconnectionDelay() {
         return reconnectionDelay;
     }
-    
+
     /**
      * The <b>eventDelayLimit</b> takes a non-negative integer representing the
      * number of milliseconds to allow the appender to block if the underlying
      * BlockingQueue is full. Once this limit is reached, the event is dropped.
-     * 
+     *
      * @param eventDelayLimit
      *            the event delay limit
      */
     public void setEventDelayLimit(Duration eventDelayLimit) {
         this.eventDelayLimit = eventDelayLimit;
     }
-    
+
     /**
      * Returns the value of the <b>eventDelayLimit</b> property.
      */
     public Duration getEventDelayLimit() {
         return eventDelayLimit;
     }
-    
+
     /**
      * Sets the timeout that controls how long we'll wait for the remote peer to
      * accept our connection attempt.
      * <p>
      * This property is configurable primarily to support instrumentation for unit testing.
-     * 
+     *
      * @param acceptConnectionTimeout
      *            timeout value in milliseconds
      */
